+++ lib/uri_to_file.dart
@@ -1,17 +1,12 @@
 import 'dart:io';
 
 import 'package:flutter/services.dart';
-import 'model/custom_exception/uri_io_exception.dart';
 
-/// Method Channel Helper
-const MethodChannel _methodChannel =
-    const MethodChannel('in.lazymanstudios.uritofile/helper');
+import 'src/uri_to_file_platform_interface.dart';
 
-/// Error code for PlatformException
-const String _URI_NOT_SUPPORTED = 'URI_NOT_SUPPORTED';
-
-/// Error code for PlatformException
-const String _IO_EXCEPTION = 'IO_EXCEPTION';
+export 'src/uri_to_file_ios.dart';
+export 'src/uri_to_file_method_channel.dart';
+export 'src/uri_to_file_platform_interface.dart';
 
 /// Check uri is supported or not
 ///
@@ -27,17 +22,7 @@ const String _IO_EXCEPTION = 'IO_EXCEPTION';
 /// - true  if uri supported
 /// - false if uri not supported
 bool isUriSupported(String uriString) {
-  Uri uri = Uri.parse(uriString);
-  if (Platform.isAndroid && uri.isScheme('content')) {
-    return true;
-  }
-
-  try {
-    File.fromUri(uri);
-    return true;
-  } catch (e) {
-    return false;
-  }
+  return UriToFilePlatform.instance.isUriSupported(uriString);
 }
 
 /// Create a [File] object from a uri.
@@ -50,34 +35,12 @@ bool isUriSupported(String uriString) {
 ///
 /// Supported uri scheme same as supported by File.fromUri(uri) with content uri (Android Only)
 ///
-/// If uri cannot reference a file this throws [UnsupportedError] or [IOException]
-Future<File> toFile(String uriString) async {
-  Uri uri = Uri.parse(uriString);
-  if (Platform.isAndroid && uri.isScheme('content')) {
-    try {
-      String filepath = await _methodChannel
-          .invokeMethod("fromUri", {"uriString": uriString});
-      return File(filepath);
-    } on PlatformException catch (e) {
-      switch (e.code) {
-        case _URI_NOT_SUPPORTED:
-          {
-            throw UnsupportedError(
-                'Cannot extract a file path from a ${uri.scheme} URI');
-          }
-        case _IO_EXCEPTION:
-          {
-            throw UriIOException(e.message);
-          }
-        default:
-          {
-            rethrow;
-          }
-      }
-    } catch (e) {
-      rethrow;
-    }
-  }
+/// If uri cannot reference a file this throws [PlatformException] or [IOException]
+Future<File> toFile(String uriString) {
+  return UriToFilePlatform.instance.toFile(uriString);
+}
 
-  return File.fromUri(uri);
+/// Clears temporary files created by this plugin
+Future<void> clearTemporaryFiles() {
+  return UriToFilePlatform.instance.clearTemporaryFiles();
 }