+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.0
+
+- Fixed error with files having no extension in their name
+- Updated package to support flutter 3.7.1
+
 ## 0.2.0
 
 - Breaking changes